requirements.txt
requests
httpx
sentence-transformers
faiss-cpu
chromadb
pdfminer.six
openai
numpy
fastapi
uvicorn
python-dotenv
